fixed class names for current branch
