+++ src/MethodsExercises.java
@@ -11,7 +11,6 @@ public class MethodsExercises {
 //        System.out.println(modulus(8, 4));
 //        getFactorial(getInteger(1, 10));
 //        diceRoll(6);
-        theGame();
     }
 
     public static int add(int x, int y){
@@ -104,28 +103,6 @@ public class MethodsExercises {
         }
     }
 
-    public static void theGame(){
-       Scanner sc = new Scanner(System.in);
-        int random = randomNumber(1, 100);
-        System.out.println("Lets play Guess That Number!");
-        System.out.println("Press ENTER to begin.");
-        sc.nextLine();
-        System.out.println("guess a number between 1 - 100");
-        System.out.println("Starto!!!");
-       while (true) {
-            int number = sc.nextInt();
-            if (number > 100 || number < 1) {
-                System.out.println("now now... between 1 - 100 plz!");
-            } else if (number > random) {
-                System.out.println("Lower!");
-            } else if (number == random) {
-                System.out.println("Stop Thats it!");
-                break;
-            } else if (number < random){
-                System.out.println("Higher!");
-            }
-        }
-    }
 }
